agent/requirements.txt
certifi==2018.10.15
cfssl==0.0.3b243
pyOpenSSL==16.2.0
requests==2.20.1
sh==1.12.14
